edited deck

+++ shared/src/main/scala/poolmate/Entity.scala
@@ -59,7 +59,7 @@ final case class Surface(id: Long = 0,
 
 final case class Deck(id: Long = 0,
                       poolId: Long = 0,
-                      installed: Int = 0,
+                      installed: Long = 0,
                       kind: String = "",
                       cost: Int = 0) extends Entity:
   def display = kind
